combining server code into server.js, and implementing save to db of cart
